+++ client/src/app/modules/thread/pages/thread-create/thread-create.component.ts
@@ -46,7 +46,9 @@ export class ThreadCreateComponent implements OnInit {
         (error) => {
           this.loading = false;
           console.log('-----> error', error);
-          this.snackBar.open(error, 'Ok');
+          this.snackBar.open(error, 'Ok', {
+            duration: 3000
+          });
         }
       );
