+++ client/src/app/modules/thread/pages/thread-detail/thread-detail.component.html
@@ -1,3 +1,7 @@
+<!-- store state here -->
+<ng-container [templateStore]="{ check: (isLoggedIn$ | async) }" #isLoggedIn="templateStore"></ng-container>
+<ng-container [templateStore]="{ user: (authUser$ | async) }" #auth="templateStore"></ng-container>
+
 <div class="row">
   <ng-container *ngIf="(thread$ | async) as thread">
     <div class="col-12 col-md-8">
@@ -8,25 +12,29 @@
           <p class="blog-post-meta">{{ thread.createdAt | amDateFormat:'LL' }} by
             <a [routerLink]="['/profile', thread.creator.fullName]">{{ thread.creator.fullName }}</a>
           </p>
-          <div>
-            <button mat-icon-button [matMenuTriggerFor]="moreOptions" aria-label="Delete thread">
-              <mat-icon>more_vert</mat-icon>
-            </button>
-            <mat-menu #moreOptions="matMenu" yPosition="below">
-              <button mat-menu-item (click)="deleteThread(thread.id)">
-                <mat-icon>delete</mat-icon>
-                <span>Delete</span>
-              </button>
-              <!-- <button mat-menu-item disabled>
-                <mat-icon>voicemail</mat-icon>
-                <span>Check voice mail</span>
-              </button>
-              <button mat-menu-item>
-                <mat-icon>notifications_off</mat-icon>
-                <span>Disable alerts</span>
-              </button> -->
-            </mat-menu>
-          </div>
+          <ng-container *ngIf="isLoggedIn.check">
+            <ng-container *ngIf="auth.user.userObject.id === thread.creator.id">
+              <div>
+                <button mat-icon-button [matMenuTriggerFor]="moreOptions" aria-label="Delete thread">
+                  <mat-icon>more_vert</mat-icon>
+                </button>
+                <mat-menu #moreOptions="matMenu" yPosition="below">
+                  <button mat-menu-item (click)="deleteThread(thread.id)">
+                    <mat-icon>delete</mat-icon>
+                    <span>Delete</span>
+                  </button>
+                  <!-- <button mat-menu-item disabled>
+                                        <mat-icon>voicemail</mat-icon>
+                                        <span>Check voice mail</span>
+                                      </button>
+                                      <button mat-menu-item>
+                                        <mat-icon>notifications_off</mat-icon>
+                                        <span>Disable alerts</span>
+                                      </button> -->
+                </mat-menu>
+              </div>
+            </ng-container>
+          </ng-container>
         </div>
         <hr>
         <p>{{ thread.body }}</p>
