+++ server/src/middlewares/authentication.js
